Modularized Objects base class and centralized astronomical computation logic.
- Hoisted `_prepare_observer` to `Objects` base class for standardized time/observer resolution.
- Implemented default `compute` method in `Objects` using vectorized geometric formulas.
- Decomposed `get_visible` into smaller helper methods to reduce cyclomatic complexity.
- Generalized `_vectorized_geometric_compute` to support non-Star objects via relaxed validity mask.
- Removed redundant `compute` implementations and observer logic from `Stars`, `Messier`, `NGC`, and `SolarObjects` subclasses.
- Updated `SolarObjects.compute` to delegate to `super().compute()`.
- Verified all unit tests pass (Stars, Messier, NGC, SolarObjects, and Events).
