+++ apts/objects/base.py
@@ -3,6 +3,7 @@ import types
 import unittest.mock
 from abc import ABC, abstractmethod
 from typing import Any, cast
+from types import SimpleNamespace
 
 import numpy as np
 import pandas as pd
@@ -26,9 +27,34 @@ class Objects(ABC):
     def get_skyfield_object(self, obj) -> object:
         pass
 
-    @abstractmethod
-    def compute(self, calculation_date=None, df_to_compute=None) -> pd.DataFrame | None:
-        pass
+    def compute(self, calculation_date=None, df_to_compute=None) -> pd.DataFrame:
+        """
+        Default compute implementation using vectorized geometric formulas.
+        Works for all catalogs (Stars, Messier, NGC, SolarObjects).
+        """
+        observer_to_use, _ = self._prepare_observer(calculation_date)
+
+        # If no specific DataFrame is provided, use the class's default.
+        if df_to_compute is None:
+            df_to_compute = self.objects
+
+        # Work on a copy to avoid modifying the original DataFrame slice
+        computed_df = df_to_compute.copy()
+
+        # Fast transit and altitude calculation
+        transits, alts, rises, sets = self._vectorized_geometric_compute(
+            computed_df, observer_to_use
+        )
+        computed_df[ObjectTableLabels.TRANSIT] = transits
+        computed_df[ObjectTableLabels.ALTITUDE] = alts
+        computed_df[ObjectTableLabels.RISING] = rises
+        computed_df[ObjectTableLabels.SETTING] = sets
+
+        # Always update the master objects DataFrame to keep data in sync.
+        # This handles both full catalog computations and subset-specific updates.
+        self.objects.update(computed_df)
+
+        return computed_df
 
     def __init__(self, place, calculation_date=None):
         self.place = place
@@ -56,6 +82,33 @@ class Objects(ABC):
         cast(Any, self.__dict__).update(state)
         # Re-create the unpicklable entries.
         self.ts = get_timescale()
+
+    def _prepare_observer(self, calculation_date):
+        """Prepares the observer and time for computation."""
+        if calculation_date is not None:
+            # It's a Skyfield Time object. If it's an array, use the first element.
+            if hasattr(calculation_date, "shape") and calculation_date.shape:
+                t = calculation_date[0]
+            elif isinstance(calculation_date, type(self.ts.now())):
+                t = calculation_date
+            else:
+                t = self.ts.utc(calculation_date)
+
+            # Avoid creating a whole new Place object, which is slow.
+            # We only need basic properties for computation.
+            observer_to_use = SimpleNamespace(
+                date=t,
+                local_timezone=self.place.local_timezone,
+                lat_decimal=self.place.lat_decimal,
+                lon_decimal=self.place.lon_decimal,
+                elevation=self.place.elevation,
+                observer=self.place.observer,
+                sun=getattr(self.place, "sun", None),
+            )
+        else:
+            observer_to_use = self.place
+            t = self.place.date
+        return observer_to_use, t
 
     def _filter_by_magnitude(
         self, conditions, star_magnitude_limit, limiting_magnitude
@@ -172,6 +225,65 @@ class Objects(ABC):
                         self.objects.loc[visible_candidate_objects.index]
                     )
 
+    def _prepare_visibility_check_times(self, start, stop):
+        """Prepares the time grid for visibility checking."""
+        ts = self.ts
+        t_start = ts.utc(start)
+        t_stop = ts.utc(stop)
+        # Use 100 points to match previous behavior and ensure precision
+        return ts.linspace(t_start, t_stop, 100)
+
+    def _perform_visibility_check(self, candidate_objects, check_times, conditions):
+        """Core visibility check logic."""
+        visible_mask = np.zeros(len(candidate_objects), dtype=bool)
+
+        # Optimization: Identify stars vs moving bodies (others).
+        # For star-only catalogs (NGC, Messier), we can bypass expensive row-wise
+        # Skyfield object instantiation and property access by using a class flag.
+        if getattr(self, "is_star_catalog", False) and "ra_hours" in candidate_objects.columns:
+            is_star = np.ones(len(candidate_objects), dtype=bool)
+            skyfield_objs = None
+        else:
+            # Traditional path for mixed or moving-body catalogs.
+            # Fast property extraction instead of iterrows()
+            if "skyfield_object" in candidate_objects.columns:
+                skyfield_objs = cast(
+                    pd.Series, candidate_objects["skyfield_object"]
+                ).to_numpy()
+            else:
+                skyfield_objs = np.array(
+                    [self.get_skyfield_object(row) for row in candidate_objects.itertuples()]
+                )
+            is_star = np.array([isinstance(obj, Star) for obj in skyfield_objs])
+
+        stars_indices = np.where(is_star)[0]
+        if skyfield_objs is not None:
+            other_indices = np.where(~is_star & pd.notnull(skyfield_objs))[0]
+        else:
+            other_indices = np.array([], dtype=int)
+
+        if len(stars_indices) > 0:
+            self._get_visible_stars(
+                candidate_objects,
+                skyfield_objs,
+                stars_indices,
+                check_times,
+                conditions,
+                visible_mask,
+            )
+
+        observer = self.place.observer
+        # Optimization: move observer.at(check_times) out of the loop
+        # to reuse calculated observer positions for all objects.
+        obs_at_check_times = observer.at(check_times)
+
+        # Check Other objects (like planets)
+        self._get_visible_other(
+            skyfield_objs, other_indices, obs_at_check_times, conditions, visible_mask
+        )
+
+        return visible_mask
+
     def get_visible(
         self,
         conditions,
@@ -193,11 +305,7 @@ class Objects(ABC):
             return pd.DataFrame(columns=self.objects.columns)
 
         # Vectorized visibility check
-        ts = self.ts
-        t_start = ts.utc(start)
-        t_stop = ts.utc(stop)
-        # Use 100 points to match previous behavior and ensure precision
-        check_times = ts.linspace(t_start, t_stop, 100)
+        check_times = self._prepare_visibility_check_times(start, stop)
 
         # Check if get_altaz_curve is mocked or overridden (common in tests)
         # Original method is a bound method, mocked/overridden is often a function or a Mock object
@@ -213,53 +321,9 @@ class Objects(ABC):
                 pd.DataFrame, candidate_objects.loc[visible_objects_indices].copy()
             )
         else:
-            visible_mask = np.zeros(len(candidate_objects), dtype=bool)
-
-            # Optimization: Identify stars vs moving bodies (others).
-            # For star-only catalogs (NGC, Messier), we can bypass expensive row-wise
-            # Skyfield object instantiation and property access by using a class flag.
-            if getattr(self, "is_star_catalog", False) and "ra_hours" in candidate_objects.columns:
-                is_star = np.ones(len(candidate_objects), dtype=bool)
-                skyfield_objs = None
-            else:
-                # Traditional path for mixed or moving-body catalogs.
-                # Fast property extraction instead of iterrows()
-                if "skyfield_object" in candidate_objects.columns:
-                    skyfield_objs = cast(
-                        pd.Series, candidate_objects["skyfield_object"]
-                    ).to_numpy()
-                else:
-                    skyfield_objs = np.array(
-                        [self.get_skyfield_object(row) for row in candidate_objects.itertuples()]
-                    )
-                is_star = np.array([isinstance(obj, Star) for obj in skyfield_objs])
-
-            stars_indices = np.where(is_star)[0]
-            if skyfield_objs is not None:
-                other_indices = np.where(~is_star & pd.notnull(skyfield_objs))[0]
-            else:
-                other_indices = np.array([], dtype=int)
-
-            if len(stars_indices) > 0:
-                self._get_visible_stars(
-                    candidate_objects,
-                    skyfield_objs,
-                    stars_indices,
-                    check_times,
-                    conditions,
-                    visible_mask,
-                )
-
-            observer = self.place.observer
-            # Optimization: move observer.at(check_times) out of the loop
-            # to reuse calculated observer positions for all objects.
-            obs_at_check_times = observer.at(check_times)
-
-            # Check Other objects (like planets)
-            self._get_visible_other(
-                skyfield_objs, other_indices, obs_at_check_times, conditions, visible_mask
+            visible_mask = self._perform_visibility_check(
+                candidate_objects, check_times, conditions
             )
-
             visible_candidate_objects: pd.DataFrame = cast(
                 pd.DataFrame, candidate_objects.loc[visible_mask].copy()
             )
@@ -301,7 +365,7 @@ class Objects(ABC):
 
     def _vectorized_geometric_compute(self, df, observer):
         """
-        Fast transit, altitude, rising, and setting calculation for a DataFrame of Stars.
+        Fast transit, altitude, rising, and setting calculation for a DataFrame of objects.
         Uses vectorized numpy operations and geometric approximations for speed.
 
         Note: Rising/setting times use a geometric formula that accounts for atmospheric
@@ -320,11 +384,12 @@ class Objects(ABC):
                 valid_mask = pd.notna(ras) & pd.notna(decs)
             elif "skyfield_object" in df.columns:
                 sky_objs = df["skyfield_object"].to_numpy()
-                valid_mask = np.array([isinstance(obj, Star) for obj in sky_objs])
+                # Relaxed validity check: any valid Skyfield object in the column
+                valid_mask = pd.notnull(sky_objs)
             else:
                 valid_mask = np.ones(len(df), dtype=bool)
         else:
-            # Fallback: Extract Skyfield Star objects and their coordinates in a vectorized way
+            # Fallback: Extract Skyfield objects and their coordinates in a vectorized way
             if "skyfield_object" in df.columns:
                 sky_objs = df["skyfield_object"].to_numpy()
             else:
@@ -333,13 +398,23 @@ class Objects(ABC):
                     [self.get_skyfield_object(row) for row in df.itertuples()]
                 )
 
-            valid_mask = np.array([isinstance(obj, Star) for obj in sky_objs])
+            valid_mask = pd.notnull(sky_objs)
 
             ras = np.array(
-                [obj.ra.hours if isinstance(obj, Star) else 0 for obj in sky_objs]
+                [
+                    getattr(obj, "ra", getattr(obj, "target_ra", None)).hours
+                    if obj is not None and hasattr(obj, "ra")
+                    else 0
+                    for obj in sky_objs
+                ]
             )
             decs = np.array(
-                [obj.dec.degrees if isinstance(obj, Star) else 0 for obj in sky_objs]
+                [
+                    getattr(obj, "dec", getattr(obj, "target_dec", None)).degrees
+                    if obj is not None and hasattr(obj, "dec")
+                    else 0
+                    for obj in sky_objs
+                ]
             )
 
         return vectorized_geometric_compute(
